refactor: DTO 클래스 패키지 구조 정리
